Build example member registration form
In order to demonstrate that Mezzanine covers the requirements in issue #14, I have
created an example member registration form.

This form demonstrates the following capabilities:
- Different field types
- Helper text
- Automatic e-mail to default recipient and to registrant (tested this by hand)
- Export of registrations in web interface or CSV format

I have implemented the form in the demo_data.json fixture. The regular way of loading it
should be sufficient here as well.

I have not resolved the following issues:
- Automatically displaying 'required' as helper text with required fields
- Implementing an Integer field type.

I feel those are not that important to fix - they are not explicit requirements in the
issue, anyway.

I have not yet implemented a CAPTCHA.