Add basic e-mail address obfuscation
Issue #49 requires e-mail address obfuscation. I have implemented this as a RICHTEXT_FILTER.

* The function website.utils.filters.obfuscate_email_addresses HTML encodes all mailto:
links and all plaintext e-mail addresses in an HTML string.

* I have added unit tests that test this behaviour.

* In settings.py, this function is added to the RICHTEXT_FILTERS setting. On all rich
text output fields, this function will be applied. The data in the database remains
unchanged.

* I have slightly changed the demo data to reflect this change. On the main website,
the page 'Word lid' now contains an e-mail address. If you view the source of the page,
you will see the e-mail address is not directly readable. Note: this will only work if
you are not logged in, as otherwise there is a hidden edit box that contains the e-mail
address in plaintext).