Add document listing functionality
I added new models and a template to implement document listing functionality. This is
important to close issues #19 and #20, and also work on issue #30.

The functionality is now offered as a subclass of Page, namely a Document Listing. It
uses the Media Library to upload and select documents for display.