Add Twitter widget to templates
In accordance with issue #7, I have added support for a Twitter widget to the
template of our website. More specifically, I took the HTML of our old website
and built a template from it.

Using the demo data, it will run the default search. To tailor it to @JongeDemocraten,
we need an API key. We can then set that key in the Settings dialogue. I have
tried this and it works.