Add robots.txt functionality
I have added the django-robots app to the requirements file.
The developer of the app is not very neat with version numbering
and such, so I felt I needed to refer to an explicit version. The
version in the PyPI did not define correct __str__ methods in
models.py, so the url admin did not render correctly. This version
does that better.

Note that the sitemap lookup is not working correctly for
department websites (still refers to main website). We may simply
fix this by omitting the Sitemap entry through the
ROBOTS_USE_SITEMAP setting. It is not required: our sitemaps are
at the default locations.

Also, having different rulesets for each department does not seem
to work properly. On the other hand, if we only use robots.txt for
technical purposes (e.g. disallowing crawlers on search results
or cache pages), we might not need this distinction.

I feel this commit resolves issue #48 in its simples interpretation.