Use site name in title tag
In order to have differnt titles for different sites, I changed
the template to use {{ page.site.name }} in the title tag. This
value is set in the Site.Sites admin panel.

I have also moved the name of the department to the back of the
title to ensure a tabbed page is recognisable (otherwise all tabs
would read 'Jonge Democraten - ...'). That we are in fact the
Jonge Democraten should be easily seen from our favicon.

This ties in to issue #54, but also SEO concerns in issue #47.