Allow embedding of locally-hosted PDFs
By allowing the object tag, we enable embedding of PDFs. However,
we limit the embedding to only PDFs which are locally hosted, for
security reasons.

It works in this commit. However, inserting an embedded PDF does
not work currently due to some weirdness in TinyMCE (it converts
the type attribute to shockwave flash).

This commit is aimed at resolving issue #23.