Fix security issue in filter_non_video_iframes
A staff user who included several iframes in a richtextfield could,
in some circumstances, prevent some of them from being stripped.

In particular, if the first iframe tag contained an illegal src value
or had contents, subsequent iframe tags would not be stripped.

This issue is resolved in this commit.