Give visual feedback when user selects document
To resolve issue #64, I added some javascript to the change_form admin
template, which is used for document listings. The javscript assigns
alt texts to all fields that can be used for thumbnails of newly added
documents. When the thumbnail is displayed, it is immediately replaced
by the alt text because the thumbnail is broken (trying to display a
PDF or .docx file as an image). The alt text presents the helpful
'Document selected'.

Furthermore, I changed document listing slightly to also allow empty
descriptions. If a description is empty, it is automatically generated
from the document name. The code for this is taken from the gallery
image save code.