Fix recreate_demodata_database.sh by using a different post_save method
There seems to be problems with loading fixtures in combination with auto created models, and the different SiteRelated models.
Changed the post_init.connect with a post_save in signals.py, and save to save(update_site=False). This fixes the NOT NULL constraints error when loading the demo_data.