Add the content to the first sidebar tab (upcoming events)
The content of the first tab is different for the main websites and
for department websites. The main website displays the upcoming
events of all departments (including national). The department
websites all display their own events and national events.

I therefore added a setting, SIDEBAR_AGENDA_SITES, which decides
what events to include in the sidebar for the given site. The
department website behaviour is default. Main website behaviour
is also available ('All sites'), and 'just this site' is also
available just in case.