Add homepage header image to blogpost detail view, part of issue #70
Always add the homepage header to the context via a context_processor.
There is no other way to add this information to the blogpost view,
without copying a lot of Mezzanine code.
Performance impact of this call on every context request seems to be insignificant,
but this will have to be monitored in production.