Fix video/PDF embedding issue
The embedding of videos and PDFs was problematic, as the approach we chose for both
turned out to be mutally exclusive (we could have one of the two but not both).

I fixed this by choosing the settings that allowed for PDF embedding and changing the
instructions we give for video embedding. In this way, we lose a little bit of
convenience for users but gain that we have both functionalities at the same time.