Remove code that sets 'unused' as value for unused default settings
This code led to an unwanted side effect: since GOOGLE_ANALYTICS_ID
was now not equal to '', the Mezzanine template concluded that we
were, in fact, using Google Analytics. The code was therefore activated
on the website. The reason we did not note this earlier is probably
because it is not activated when the user is logged in.