Fix header image scaling and size
Use the mezzanine thumbnail tag to resize header images
to a width of 2500 pixels and a jpeg quality of 75.

Images are automatically resized, but only if the size does not yet exist.

This fixes gray areas on the right and left of the header on
large/high-res screens. And fixes possible large image file sizes.

Typical header image size will now be 100-300 kB.